# internal/apispec/openapi.yaml
openapi: 3.0.3
info:
  title: Community Waste Collection API
  description: |
    A RESTful API service for managing community household waste collection,
    pickup scheduling, and payment processing.

    Built with Go 1.26, Echo v4, PostgreSQL 17, and Docker.

    ## Business Rules
    - **BR-01**: A new pickup cannot be created if the household has a pending payment (409)
    - **BR-02**: A pickup can only be scheduled when its status is `pending` (409)
    - **BR-03**: An electronic pickup requires `safety_check: true` before scheduling (422)
    - **BR-04**: Organic pickups are automatically canceled after 3 days if still pending
    - **BR-05**: Completing a pickup auto-creates a payment (organic/plastic/paper: Rp 50,000; electronic: Rp 100,000)
    - **BR-06**: Payment confirmation requires a proof-of-payment file upload (stored in S3)
  version: "0.1.0"
  contact:
    name: Community Waste Collection API

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Households
    description: Household registration and management
  - name: Pickups
    description: Waste pickup requests and lifecycle management
  - name: Payments
    description: Payment creation and confirmation
  - name: Reports
    description: Aggregated reporting and household history

paths:
  /api/households:
    post:
      tags: [Households]
      summary: Create a household
      operationId: createHousehold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseholdRequest'
            example:
              owner_name: "Budi Santoso"
              address: "Jl. Merdeka No. 5, Jakarta"
      responses:
        '201':
          description: Household created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Households]
      summary: List households
      operationId: listHouseholds
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated list of households
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdListResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/households/{id}:
    get:
      tags: [Households]
      summary: Get a household by ID
      operationId: getHousehold
      parameters:
        - $ref: '#/components/parameters/HouseholdID'
      responses:
        '200':
          description: Household details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Households]
      summary: Delete a household
      operationId: deleteHousehold
      parameters:
        - $ref: '#/components/parameters/HouseholdID'
      responses:
        '204':
          description: Household deleted (cascades to pickups and payments)
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/pickups:
    post:
      tags: [Pickups]
      summary: Create a waste pickup request
      operationId: createPickup
      description: |
        Creates a new pickup request. Fails with 409 if the household has a pending
        payment (BR-01). Subject to per-IP rate limiting (429 if exceeded).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePickupRequest'
            examples:
              organic:
                summary: Organic waste pickup
                value:
                  household_id: "550e8400-e29b-41d4-a716-446655440000"
                  type: "organic"
              electronic:
                summary: Electronic waste with safety check
                value:
                  household_id: "550e8400-e29b-41d4-a716-446655440000"
                  type: "electronic"
                  safety_check: true
      responses:
        '201':
          description: Pickup created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PickupResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Pickups]
      summary: List pickup requests
      operationId: listPickups
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: household_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by household ID
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PickupStatus'
          description: Filter by status
      responses:
        '200':
          description: Paginated list of pickups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PickupListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/pickups/{id}/schedule:
    put:
      tags: [Pickups]
      summary: Schedule a pickup
      operationId: schedulePickup
      description: |
        Sets the pickup date and moves status from `pending` to `scheduled`.
        Requires status=pending (BR-02). Electronic pickups require safety_check=true (BR-03).
      parameters:
        - $ref: '#/components/parameters/PickupID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulePickupRequest'
            example:
              pickup_date: "2026-06-10T09:00:00Z"
              safety_check: true
      responses:
        '200':
          description: Pickup scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PickupResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/pickups/{id}/complete:
    put:
      tags: [Pickups]
      summary: Complete a pickup
      operationId: completePickup
      description: |
        Marks pickup as completed and automatically creates a payment (BR-05).
        Payment amount: organic/plastic/paper = Rp 50,000; electronic = Rp 100,000.
      parameters:
        - $ref: '#/components/parameters/PickupID'
      responses:
        '200':
          description: Pickup completed and payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PickupResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/pickups/{id}/cancel:
    put:
      tags: [Pickups]
      summary: Cancel a pickup
      operationId: cancelPickup
      parameters:
        - $ref: '#/components/parameters/PickupID'
      responses:
        '200':
          description: Pickup canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PickupResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/payments:
    post:
      tags: [Payments]
      summary: Create a payment
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
            example:
              household_id: "550e8400-e29b-41d4-a716-446655440000"
              waste_id: "660e8400-e29b-41d4-a716-446655440001"
              amount: "50000.00"
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Payments]
      summary: List payments
      operationId: listPayments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: household_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by household ID
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
          description: Filter by payment status
        - name: date_from
          in: query
          schema:
            type: string
            format: date-time
          description: Filter payments from this date (RFC3339)
        - name: date_to
          in: query
          schema:
            type: string
            format: date-time
          description: Filter payments up to this date (RFC3339)
      responses:
        '200':
          description: Paginated list of payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/payments/{id}/confirm:
    put:
      tags: [Payments]
      summary: Confirm a payment with proof of payment
      operationId: confirmPayment
      description: |
        Accepts a multipart form upload with the proof-of-payment file.
        Uploads the file to S3 and saves the URL to the payment record (BR-06).
      parameters:
        - $ref: '#/components/parameters/PaymentID'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [proof]
              properties:
                proof:
                  type: string
                  format: binary
                  description: Proof of payment file (image or PDF, max 10 MB)
      responses:
        '200':
          description: Payment confirmed with proof URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/reports/waste-summary:
    get:
      tags: [Reports]
      summary: Get waste pickup summary
      operationId: wasteSummary
      description: Returns aggregated pickup counts grouped by waste type and status.
      responses:
        '200':
          description: Waste summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WasteSummaryResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/reports/payment-summary:
    get:
      tags: [Reports]
      summary: Get payment summary
      operationId: paymentSummary
      description: Returns total payment counts by status and total revenue from paid payments.
      responses:
        '200':
          description: Payment summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSummaryResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/reports/households/{id}/history:
    get:
      tags: [Reports]
      summary: Get household full history
      operationId: householdHistory
      parameters:
        - $ref: '#/components/parameters/HouseholdID'
      responses:
        '200':
          description: Full pickup and payment history for the household
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdHistoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: healthCheck
      description: Returns 200 when the application and database are reachable.
      responses:
        '200':
          description: Application is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        '503':
          description: Application is unhealthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "unhealthy"
                  error:
                    type: string
                    example: "database unreachable"

components:
  parameters:
    HouseholdID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Household UUID

    PickupID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Pickup UUID

    PaymentID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Payment UUID

    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number

    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page (capped at 100)

  schemas:
    # Enums
    PickupType:
      type: string
      enum: [organic, plastic, paper, electronic]

    PickupStatus:
      type: string
      enum: [pending, scheduled, completed, canceled]

    PaymentStatus:
      type: string
      enum: [pending, paid, failed]

    # Household
    Household:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_name:
          type: string
        address:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateHouseholdRequest:
      type: object
      required: [owner_name, address]
      properties:
        owner_name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1

    HouseholdResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Household'

    HouseholdListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Household'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Pickup
    WastePickup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/PickupType'
        status:
          $ref: '#/components/schemas/PickupStatus'
        pickup_date:
          type: string
          format: date-time
          nullable: true
        safety_check:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreatePickupRequest:
      type: object
      required: [household_id, type]
      properties:
        household_id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/PickupType'
        safety_check:
          type: boolean
          default: false
          description: Required to be true for electronic pickups before scheduling

    SchedulePickupRequest:
      type: object
      required: [pickup_date]
      properties:
        pickup_date:
          type: string
          format: date-time
          description: Must be a future date/time
        safety_check:
          type: boolean
          description: Required to be true for electronic pickup types

    PickupResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/WastePickup'

    PickupListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/WastePickup'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Payment
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        waste_id:
          type: string
          format: uuid
        amount:
          type: string
          description: Decimal string (e.g. "50000.00")
          example: "50000.00"
        status:
          $ref: '#/components/schemas/PaymentStatus'
        payment_date:
          type: string
          format: date-time
          nullable: true
        proof_file_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreatePaymentRequest:
      type: object
      required: [household_id, waste_id, amount]
      properties:
        household_id:
          type: string
          format: uuid
        waste_id:
          type: string
          format: uuid
        amount:
          type: string
          description: Positive decimal string (e.g. "50000.00")

    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Payment'

    PaymentListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Reports
    PickupStatusBreakdown:
      type: object
      additionalProperties:
        type: integer
      description: Map of status → count
      example:
        pending: 5
        scheduled: 2
        completed: 4
        canceled: 1

    WasteTypeSummary:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/PickupType'
        total:
          type: integer
        by_status:
          $ref: '#/components/schemas/PickupStatusBreakdown'

    WasteSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            by_type:
              type: array
              items:
                $ref: '#/components/schemas/WasteTypeSummary'

    PaymentStatusSummary:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/PaymentStatus'
        count:
          type: integer
        total_amount:
          type: string
          description: Decimal string

    PaymentSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            by_status:
              type: array
              items:
                $ref: '#/components/schemas/PaymentStatusSummary'
            total_revenue:
              type: string
              description: Total amount from paid payments (decimal string)
              example: "150000.00"

    HouseholdHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            household:
              $ref: '#/components/schemas/Household'
            pickups:
              type: array
              items:
                $ref: '#/components/schemas/WastePickup'
            payments:
              type: array
              items:
                $ref: '#/components/schemas/Payment'

    # Shared
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: "owner_name is required"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: "household not found"

    Conflict:
      description: Business state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: CONFLICT
              message: "household has a pending payment"

    BusinessRuleViolation:
      description: Business rule violated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: BUSINESS_RULE_VIOLATION
              message: "electronic pickup requires safety_check to be true before scheduling"

    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_LIMITED
              message: "rate limit exceeded"

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: "internal server error"